+++ neuracore/__init__.py
@@ -1,5 +1,5 @@
 from .api import *  # noqa: F403
 from .core.exceptions import *  # noqa: F403
 
-__version__ = "1.2.2"
+__version__ = "1.3.1"